# psd-render-api/openapi-current.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSD 渲染服务平台 API",
    "description": "## 概述\nPSD 模板渲染服务平台对外提供 PSD 模板管理、资产上传、渲染任务提交与查询的 HTTP API。\n\n## 鉴权\n所有 `/v1/*` 接口需在 `Authorization` 头携带 API Key（格式：`Bearer sk_live_xxx`）。\nAPI Key 通过 Admin 控制台创建，每个 Key 可配置：\n- 限流（rateLimitPerMin）：每分钟最大请求数\n- 日配额（quotaPerDay）：每天最大请求数\n- 作用域（scopes）：限定可调用的接口\n- IP 白名单（ipWhitelist）\n- 默认 Webhook URL / Secret\n\n## 幂等性\n`POST /v1/render-jobs` 支持通过 `Idempotency-Key` 头实现幂等：\n同租户同 key 重复提交将返回相同任务，不重复创建。\n\n## Webhook\n任务终态（succeeded / failed / cancelled）时自动投递 Webhook 到提交时声明的 `webhookUrl`。\n投递包含 HMAC-SHA256 签名头 `X-Render-Signature`，可校验来源。\n失败自动重试，指数退避：1m / 5m / 15m / 60m / 240m / 960m。\n\n## 错误响应\n所有错误响应统一为 `{ \"error\": \"<CODE>\", \"message\": \"<描述>\", \"details\"?: [...] }`。",
    "version": "0.3.0",
    "contact": {
      "name": "PSD 渲染服务平台"
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API Key（格式：sk_live_xxx）",
        "bearerFormat": "sk_live_xxx"
      },
      "adminSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin Session Token（登录后获得）",
        "bearerFormat": "session_token"
      },
      "workerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Worker 访问令牌（注册后获得）",
        "bearerFormat": "wkr_xxx"
      },
      "storageSignedToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "存储签名令牌（格式：Bearer <token>，由 generateUploadUrl/generateDownloadUrl 签发，仅 local 模式可用。也支持 ?token= 查询参数，但不推荐）"
      },
      "workerRegisterSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Worker 注册凭据（格式：Bearer <WORKER_REGISTER_SECRET> 或一次性配对码/token，也可通过 X-Worker-Register-Secret 头传递）"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误描述"
          },
          "details": {
            "type": "array",
            "description": "可选，详细错误信息"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse"
      }
    }
  },
  "paths": {
...
    "/v1/render-jobs": {
      "post": {
        "summary": "提交渲染任务",
        "tags": [
          "render-jobs"
        ],
        "description": "提交渲染任务。任务入队后由空闲 Worker 领取执行，终态时回调 webhookUrl（若提供）。\n\n**幂等性**：通过 `Idempotency-Key` 请求头实现。同租户同 key 重复提交将返回相同任务，HTTP 200；首次创建返回 201。\n\n**能力路由**：可选的 `requiredCapabilities` 限定任务的 Worker 要求（如 PS 大版本、是否支持智能对象）。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "templateVersionId",
                  "input"
                ],
                "properties": {
                  "templateVersionId": {
                    "type": "string",
                    "description": "已发布的模板版本 ID（tpv_xxx）"
                  },
                  "input": {
                    "type": "object",
                    "description": "绑定 ID 到输入值的映射。图片绑定用 `{ assetId }`，文本绑定用 `{ text }`。",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "assetId": {
                          "type": "string",
                          "description": "上传资产后获得的 assetId（art_xxx）"
                        },
                        "text": {
                          "type": "string",
                          "description": "文本内容（仅文本绑定）"
                        }
                      }
                    }
                  },
                  "output": {
                    "type": "object",
                    "properties": {
                      "format": {
                        "type": "string",
                        "enum": [
                          "png",
                          "jpeg",
                          "psd"
                        ],
                        "default": "png"
                      },
                      "quality": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 100
                      }
                    },
                    "default": {
                      "format": "png"
                    }
                  },
                  "priority": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10,
                    "description": "优先级（数字越小优先级越高，默认 5）"
                  },
                  "webhookUrl": {
                    "type": "string",
                    "format": "uri",
                    "description": "可选，任务终态时回调此地址"
                  },
                  "requiredCapabilities": {
                    "type": "object",
                    "description": "可选，能力路由要求",
                    "properties": {
                      "psMajorVersion": {
                        "type": "integer",
                        "minimum": 1
                      },
                      "supportsSmartObject": {
                        "type": "boolean"
                      },
                      "supportsTextLayer": {
                        "type": "boolean"
                      },
                      "os": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 128
            },
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "description": "幂等键（同租户下唯一，≤128 字符）。未提供时自动生成 `auto-{timestamp}`。"
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "幂等命中时返回已有任务",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "幂等命中时返回已有任务",
                  "required": [
                    "jobId",
                    "status",
                    "idempotencyKey",
                    "traceId",
                    "createdAt"
                  ],
                  "properties": {
                    "jobId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "idempotencyKey": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "jobId",
                    "status",
                    "idempotencyKey",
                    "traceId",
                    "createdAt"
                  ],
                  "properties": {
                    "jobId": {
                      "type": "string",
                      "description": "任务编码（job_xxx）"
                    },
                    "status": {
                      "type": "string",
                      "description": "初始状态（QUEUED）"
                    },
                    "idempotencyKey": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "链路追踪 ID"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺失"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "429": {
            "description": "触发限流或日配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流或日配额耗尽"
                }
              }
            }
          }
        }
      }
    },
    "/v1/render-jobs/{jobId}": {
      "get": {
        "summary": "查询任务",
        "tags": [
          "render-jobs"
        ],
        "description": "查询任务状态、进度、产物下载地址。仅返回同租户下的任务。",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "jobId",
            "required": true,
            "description": "任务编码（job_xxx）"
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "jobId",
                    "status"
                  ],
                  "properties": {
                    "jobId": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "description": "QUEUED | LEASED | PROCESSING | SUCCEEDED | FAILED | CANCELLING | CANCELLED"
                    },
                    "progress": {
                      "type": "integer"
                    },
                    "stage": {
                      "type": "string"
                    },
                    "resultUrl": {
                      "type": "string",
                      "description": "结果下载地址（SUCCEEDED 时返回，3 天有效）"
                    },
                    "errorCode": {
                      "type": "string"
                    },
                    "errorMessage": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺失"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "429": {
            "description": "触发限流或日配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流或日配额耗尽"
                }
              }
            }
          }
        }
      }
    },
    "/v1/render-jobs/{jobId}/cancel": {
      "post": {
        "summary": "取消任务",
        "tags": [
          "render-jobs"
        ],
        "description": "请求取消任务。\n\n状态机：\n- QUEUED：直接置为 CANCELLED\n- LEASED/PROCESSING：置为 CANCELLING，Worker 在阶段边界检测后置为 CANCELLED\n- SUCCEEDED/FAILED/CANCELLED：返回 409 拒绝",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "可选，取消原因"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "jobId",
            "required": true,
            "description": "任务编码（job_xxx）"
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "jobId",
                    "status",
                    "updated",
                    "message"
                  ],
                  "properties": {
                    "jobId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "updated": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺失"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "jobId",
                    "status",
                    "updated",
                    "message"
                  ],
                  "properties": {
                    "jobId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "updated": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "触发限流或日配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流或日配额耗尽"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}